Rewrite graph comparison tool and update brute_force to comply with other algorithms

+++ create_distance_matrix.py
@@ -4,7 +4,7 @@ import numpy as np
 def create_distance_matrix(problem):
     nodes = list(problem.get_nodes())
     num_nodes = problem.dimension
-    distance_matrix = np.zeros((num_nodes, num_nodes), dtype=int)
+    distance_matrix = np.zeros((num_nodes, num_nodes), dtype=float)
     for i in range(num_nodes):
         for j in range(num_nodes):
             if i != j:
